# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outlier AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/signals": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "List Signals",
        "operationId": "list_signals_api_signals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "trajectory_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trajectory Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response List Signals Api Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Create Signal",
        "operationId": "create_signal_api_signals_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/queue": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Signals Queue",
        "operationId": "signals_queue_api_signals_queue_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response Signals Queue Api Signals Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/{signal_id}": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Get Signal",
        "operationId": "get_signal_api_signals__signal_id__get",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/{signal_id}/confirm": {
      "put": {
        "tags": [
          "signals"
        ],
        "summary": "Confirm Signal",
        "operationId": "confirm_signal_api_signals__signal_id__confirm_put",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/{signal_id}/reject": {
      "put": {
        "tags": [
          "signals"
        ],
        "summary": "Reject Signal",
        "operationId": "reject_signal_api_signals__signal_id__reject_put",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals/extract": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Extract Signals",
        "description": "Group this trajectory's stored comments into proposed signals (theme, count, evidence).",
        "operationId": "extract_signals_api_signals_extract_post",
        "parameters": [
          {
            "name": "trajectory_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trajectory Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalOut"
                  },
                  "title": "Response Extract Signals Api Signals Extract Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Sentiment": {
        "type": "string",
        "enum": [
          "neg",
          "neu",
          "pos"
        ],
        "title": "Sentiment"
      },
      "SignalCreate": {
        "properties": {
          "trajectory_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trajectory Id"
          },
          "render_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Id"
          },
          "kind": {
            "$ref": "#/components/schemas/SignalKind",
            "default": "note"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "sentiment": {
            "$ref": "#/components/schemas/Sentiment",
            "default": "neu"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "trajectory_id",
          "text"
        ],
        "title": "SignalCreate"
      },
      "SignalKind": {
        "type": "string",
        "enum": [
          "comment_theme",
          "objection",
          "praise",
          "joke",
          "misreading",
          "question",
          "competitor_mention",
          "quote",
          "hook_stat",
          "share_pattern",
          "note"
        ],
        "title": "SignalKind"
      },
      "SignalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trajectory_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trajectory Id"
          },
          "render_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "extracted_by": {
            "type": "string",
            "title": "Extracted By"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trajectory_id",
          "render_id",
          "kind",
          "text",
          "evidence",
          "sentiment",
          "count",
          "extracted_by",
          "status",
          "decided_by",
          "decided_at",
          "created_at"
        ],
        "title": "SignalOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}